+++ xx-economics/src/weights.rs
@@ -57,22 +57,22 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: XXEconomics InflationParams (r:0 w:1)
 	fn set_inflation_params() -> Weight {
 		Weight::from_parts(18_743_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics InterestPoints (r:0 w:1)
 	fn set_interest_points() -> Weight {
 		Weight::from_parts(19_173_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics IdealLiquidityStake (r:0 w:1)
 	fn set_liquidity_rewards_stake() -> Weight {
 		Weight::from_parts(18_771_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics LiquidityRewards (r:0 w:1)
 	fn set_liquidity_rewards_balance() -> Weight {
 		Weight::from_parts(18_888_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 }
 
@@ -81,21 +81,21 @@ impl WeightInfo for () {
 	// Storage: XXEconomics InflationParams (r:0 w:1)
 	fn set_inflation_params() -> Weight {
 		Weight::from_parts(18_743_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics InterestPoints (r:0 w:1)
 	fn set_interest_points() -> Weight {
 		Weight::from_parts(19_173_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics IdealLiquidityStake (r:0 w:1)
 	fn set_liquidity_rewards_stake() -> Weight {
 		Weight::from_parts(18_771_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: XXEconomics LiquidityRewards (r:0 w:1)
 	fn set_liquidity_rewards_balance() -> Weight {
 		Weight::from_parts(18_888_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 }
