+++ xx-team-custody/src/weights.rs
@@ -64,8 +64,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: XXCustody TeamAccounts (r:1 w:1)
 	fn payout() -> Weight {
 		Weight::from_parts(127_699_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(5 as u64))
-			.saturating_add(T::DbWeight::get().writes(5 as u64))
+			.saturating_add(T::DbWeight::get().reads(5_u64))
+			.saturating_add(T::DbWeight::get().writes(5_u64))
 	}
 	// Storage: System Account (r:1 w:1)
 	// Storage: Staking HistoryDepth (r:1 w:0)
@@ -77,8 +77,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: Balances Locks (r:1 w:1)
 	fn custody_bond() -> Weight {
 		Weight::from_parts(94_786_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(8 as u64))
-			.saturating_add(T::DbWeight::get().writes(4 as u64))
+			.saturating_add(T::DbWeight::get().reads(8_u64))
+			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
 	// Storage: Balances Locks (r:1 w:1)
 	// Storage: XXCustody Custodians (r:1 w:0)
@@ -88,8 +88,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: System Account (r:1 w:1)
 	fn custody_bond_extra() -> Weight {
 		Weight::from_parts(77_928_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(6 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(6_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: Staking Bonded (r:1 w:1)
 	// Storage: XXCustody Custodians (r:1 w:0)
@@ -97,8 +97,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: XXCustody CustodyAccounts (r:1 w:0)
 	fn custody_set_controller() -> Weight {
 		Weight::from_parts(45_799_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(5 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(5_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: System Account (r:1 w:1)
 	// Storage: Proxy Proxies (r:1 w:1)
@@ -106,32 +106,32 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: XXCustody Custodians (r:1 w:0)
 	fn custody_set_proxy() -> Weight {
 		Weight::from_parts(54_788_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
 	// Storage: Proxy Proxies (r:1 w:1)
 	// Storage: System Account (r:2 w:2)
 	// Storage: XXCustody TeamAccounts (r:1 w:1)
 	fn team_custody_set_proxy() -> Weight {
 		Weight::from_parts(89_382_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(4 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(4_u64))
 	}
 	// Storage: XXCustody Custodians (r:0 w:1)
 	fn add_custodian() -> Weight {
 		Weight::from_parts(20_959_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: XXCustody Custodians (r:0 w:1)
 	fn remove_custodian() -> Weight {
 		Weight::from_parts(21_083_000, 1000)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(T::DbWeight::get().writes(1_u64))
 	}
 	// Storage: XXCustody TeamAccounts (r:2 w:2)
 	fn replace_team_member() -> Weight {
 		Weight::from_parts(36_760_000, 1000)
-			.saturating_add(T::DbWeight::get().reads(2 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(T::DbWeight::get().reads(2_u64))
+			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
 }
 
@@ -142,8 +142,8 @@ impl WeightInfo for () {
 	// Storage: XXCustody TeamAccounts (r:1 w:1)
 	fn payout() -> Weight {
 		Weight::from_parts(127_699_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(5 as u64))
-			.saturating_add(RocksDbWeight::get().writes(5 as u64))
+			.saturating_add(RocksDbWeight::get().reads(5_u64))
+			.saturating_add(RocksDbWeight::get().writes(5_u64))
 	}
 	// Storage: System Account (r:1 w:1)
 	// Storage: Staking HistoryDepth (r:1 w:0)
@@ -155,8 +155,8 @@ impl WeightInfo for () {
 	// Storage: Balances Locks (r:1 w:1)
 	fn custody_bond() -> Weight {
 		Weight::from_parts(94_786_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(8 as u64))
-			.saturating_add(RocksDbWeight::get().writes(4 as u64))
+			.saturating_add(RocksDbWeight::get().reads(8_u64))
+			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
 	// Storage: Balances Locks (r:1 w:1)
 	// Storage: XXCustody Custodians (r:1 w:0)
@@ -166,8 +166,8 @@ impl WeightInfo for () {
 	// Storage: System Account (r:1 w:1)
 	fn custody_bond_extra() -> Weight {
 		Weight::from_parts(77_928_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(6 as u64))
-			.saturating_add(RocksDbWeight::get().writes(3 as u64))
+			.saturating_add(RocksDbWeight::get().reads(6_u64))
+			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
 	// Storage: Staking Bonded (r:1 w:1)
 	// Storage: XXCustody Custodians (r:1 w:0)
@@ -175,8 +175,8 @@ impl WeightInfo for () {
 	// Storage: XXCustody CustodyAccounts (r:1 w:0)
 	fn custody_set_controller() -> Weight {
 		Weight::from_parts(45_799_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(5 as u64))
-			.saturating_add(RocksDbWeight::get().writes(3 as u64))
+			.saturating_add(RocksDbWeight::get().reads(5_u64))
+			.saturating_add(RocksDbWeight::get().writes(3_u64))
 	}
 	// Storage: System Account (r:1 w:1)
 	// Storage: Proxy Proxies (r:1 w:1)
@@ -184,31 +184,31 @@ impl WeightInfo for () {
 	// Storage: XXCustody Custodians (r:1 w:0)
 	fn custody_set_proxy() -> Weight {
 		Weight::from_parts(54_788_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(4 as u64))
-			.saturating_add(RocksDbWeight::get().writes(2 as u64))
+			.saturating_add(RocksDbWeight::get().reads(4_u64))
+			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
 	// Storage: Proxy Proxies (r:1 w:1)
 	// Storage: System Account (r:2 w:2)
 	// Storage: XXCustody TeamAccounts (r:1 w:1)
 	fn team_custody_set_proxy() -> Weight {
 		Weight::from_parts(89_382_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(4 as u64))
-			.saturating_add(RocksDbWeight::get().writes(4 as u64))
+			.saturating_add(RocksDbWeight::get().reads(4_u64))
+			.saturating_add(RocksDbWeight::get().writes(4_u64))
 	}
 	// Storage: XXCustody Custodians (r:0 w:1)
 	fn add_custodian() -> Weight {
 		Weight::from_parts(20_959_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: XXCustody Custodians (r:0 w:1)
 	fn remove_custodian() -> Weight {
 		Weight::from_parts(21_083_000, 1000)
-			.saturating_add(RocksDbWeight::get().writes(1 as u64))
+			.saturating_add(RocksDbWeight::get().writes(1_u64))
 	}
 	// Storage: XXCustody TeamAccounts (r:2 w:2)
 	fn replace_team_member() -> Weight {
 		Weight::from_parts(36_760_000, 1000)
-			.saturating_add(RocksDbWeight::get().reads(2 as u64))
-			.saturating_add(RocksDbWeight::get().writes(2 as u64))
+			.saturating_add(RocksDbWeight::get().reads(2_u64))
+			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
 }
